organize JavaScript with modules
