+++ js/index.js
@@ -1,0 +1,3 @@
+import startApp from '../modules/startApp.js';
+
+startApp();
