+++ modules/store.js
@@ -1,28 +1,4 @@
-class Book {
-  constructor({ author, title, pages, read = false, id = this.generateID() }) {
-    if (Book.isValid({ author, title, pages })) {
-      this.author = author;
-      this.title = title;
-      this.pages = pages;
-      this.read = read;
-      this.id = id;
-    }
-  }
-
-  static isValid = ({ author, title, pages }) => typeof author === 'string'
-    && author.length > 0
-    && typeof title === 'string'
-    && title.length > 0
-    && typeof pages === 'number';
-
-  generateID = () => `_${Math.random()
-    .toString(36)
-    .substr(2, 9)}`;
-
-  toggleRead = () => {
-    this.read = !this.read;
-  };
-}
+import Book from './book-class.js';
 
 class Store {
   constructor(initialData = []) {
@@ -34,14 +10,14 @@ class Store {
     }
   }
 
-  static setToLocalStorage = item => {
+  static #setToLocalStorage = item => {
     localStorage.setItem('lanr3wajuAwesomeBooksLibrary', JSON.stringify(item));
   };
 
   add = (book = {}) => {
     this.newBook = new Book(book);
     this.memory = [...this.memory, this.newBook];
-    Store.setToLocalStorage(this.memory);
+    Store.#setToLocalStorage(this.memory);
     return this.newBook;
   };
 
@@ -52,13 +28,13 @@ class Store {
       }
       return element;
     });
-    Store.setToLocalStorage(this.memory);
+    Store.#setToLocalStorage(this.memory);
     return this.memory;
   };
 
   remove = id => {
     this.memory = this.memory.filter(book => book.id !== id);
-    Store.setToLocalStorage(this.memory);
+    Store.#setToLocalStorage(this.memory);
     return this.memory;
   };
